refactor(spirv): move implementation to internal/codegen — DXIL pattern (ARCH-001)
All codegen moved to spirv/internal/codegen/ via git mv. Thin public
wrapper at spirv/spirv.go with type aliases. Public API unchanged:
Backend, Options, DefaultOptions, ModuleBuilder, capabilities, opcodes.
SPIR-V val 172/172 (100%). 0 lint issues, all tests pass.
