+++ spirv/internal/codegen/atomic_type_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"testing"
