+++ spirv/internal/codegen/backend.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"fmt"
