+++ spirv/internal/codegen/bench_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"runtime"
