+++ spirv/internal/codegen/capabilities_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"encoding/binary"
