+++ spirv/internal/codegen/image_bounds_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"testing"
