+++ spirv/internal/codegen/phase3_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"testing"
