+++ spirv/internal/codegen/ptr_math_gather_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"strings"
