+++ spirv/internal/codegen/ray_query.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"fmt"
