+++ spirv/internal/codegen/sdf_analysis_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"encoding/binary"
