+++ spirv/internal/codegen/vec2_ifelse_cache_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"os"
