+++ spirv/internal/codegen/vello_workaround_test.go
@@ -1,4 +1,4 @@
-package spirv
+package codegen
 
 import (
 	"os"
