+++ spirv/spirv.go
@@ -1,906 +1,822 @@
 // Package spirv provides SPIR-V code generation from naga IR.
 //
-// SPIR-V is the standard intermediate language for GPU shaders,
-// used by Vulkan, OpenCL, and other APIs.
+// This package exposes the public API for SPIR-V compilation.
+// All implementation details live in spirv/internal/codegen.
 package spirv
 
-import "github.com/gogpu/naga/ir"
+import (
+	"github.com/gogpu/naga/ir"
+	"github.com/gogpu/naga/spirv/internal/codegen"
+)
+
+// --- Configuration types ---
 
 // Version represents a SPIR-V version.
-type Version struct {
-	Major uint8
-	Minor uint8
-}
+type Version = codegen.Version
 
-// Common SPIR-V versions
+// Common SPIR-V versions.
 var (
-	Version1_0 = Version{1, 0}
-	Version1_1 = Version{1, 1}
-	Version1_2 = Version{1, 2}
-	Version1_3 = Version{1, 3}
-	Version1_4 = Version{1, 4}
-	Version1_5 = Version{1, 5}
-	Version1_6 = Version{1, 6}
+	Version1_0 = codegen.Version1_0
+	Version1_1 = codegen.Version1_1
+	Version1_2 = codegen.Version1_2
+	Version1_3 = codegen.Version1_3
+	Version1_4 = codegen.Version1_4
+	Version1_5 = codegen.Version1_5
+	Version1_6 = codegen.Version1_6
 )
 
 // Options configures SPIR-V generation.
-type Options struct {
-	// Version is the SPIR-V version to target
-	Version Version
-
-	// Capabilities are additional capabilities to declare
-	Capabilities []Capability
-
-	// Debug includes debug information
-	Debug bool
-
-	// Validation enables output validation
-	Validation bool
-
-	// UseStorageInputOutput16 enables StorageInputOutput16 capability for f16
-	// types used in entry point inputs/outputs. When true, f16 I/O uses native
-	// 16-bit types; when false, a polyfill converts to/from f32.
-	UseStorageInputOutput16 bool
-
-	// ForcePointSize adds a BuiltIn PointSize output variable with value 1.0
-	// to every vertex shader entry point. Required by some Vulkan drivers.
-	ForcePointSize bool
-
-	// AdjustCoordinateSpace flips the Y coordinate of Position built-in outputs
-	// in vertex shaders to match Vulkan's coordinate system (Y pointing down).
-	AdjustCoordinateSpace bool
-
-	// ForceLoopBounding inserts a decrementing counter at every loop header
-	// that breaks when the counter reaches zero. This prevents infinite loops
-	// from hanging the GPU. Matches Rust naga's force_loop_bounding option
-	// (default true). The counter uses a vec2<u32> to simulate 64-bit range.
-	ForceLoopBounding bool
-
-	// BoundsCheckPolicies controls how out-of-bounds image accesses are handled.
-	BoundsCheckPolicies BoundsCheckPolicies
-
-	// CapabilitiesAvailable limits which capabilities may be used. When nil,
-	// all capabilities are available (the default). When non-nil, only
-	// capabilities present in the set may be used; others trigger polyfills.
-	// This matches Rust naga's Options::capabilities field.
-	CapabilitiesAvailable map[Capability]struct{}
-
-	// RayQueryInitTracking enables initialization tracking for ray queries.
-	// When true (default), ray query helper functions include validation checks
-	// that track whether the query was properly initialized before proceed/get.
-	// When false, validation checks are skipped and helper functions branch
-	// unconditionally. Matches Rust naga's ray_query_initialization_tracking.
-	RayQueryInitTracking bool
-}
+type Options = codegen.Options
 
 // BoundsCheckPolicy controls how out-of-bounds resource accesses are handled.
-type BoundsCheckPolicy uint8
+type BoundsCheckPolicy = codegen.BoundsCheckPolicy
 
+// BoundsCheckPolicy values.
 const (
-	// BoundsCheckUnchecked performs no bounds checking (default).
-	BoundsCheckUnchecked BoundsCheckPolicy = iota
-	// BoundsCheckRestrict clamps coordinates/level/sample to valid range.
-	BoundsCheckRestrict
-	// BoundsCheckReadZeroSkipWrite returns zero for out-of-bounds reads, skips writes.
-	BoundsCheckReadZeroSkipWrite
+	BoundsCheckUnchecked         = codegen.BoundsCheckUnchecked
+	BoundsCheckRestrict          = codegen.BoundsCheckRestrict
+	BoundsCheckReadZeroSkipWrite = codegen.BoundsCheckReadZeroSkipWrite
 )
 
 // BoundsCheckPolicies holds per-resource-type bounds check policies.
-type BoundsCheckPolicies struct {
-	// ImageLoad controls bounds checking for image load operations.
-	ImageLoad BoundsCheckPolicy
-	// ImageStore controls bounds checking for image store operations.
-	ImageStore BoundsCheckPolicy
-	// Index controls bounds checking for buffer index operations.
-	Index BoundsCheckPolicy
-}
+type BoundsCheckPolicies = codegen.BoundsCheckPolicies
 
 // DefaultOptions returns sensible default options.
 func DefaultOptions() Options {
-	return Options{
-		Version:                 Version1_1,
-		Debug:                   false,
-		Validation:              true,
-		UseStorageInputOutput16: true,
-		ForceLoopBounding:       true,
-		RayQueryInitTracking:    true,
-	}
+	return codegen.DefaultOptions()
 }
+
+// --- Capability ---
 
 // Capability represents a SPIR-V capability.
-type Capability uint32
+type Capability = codegen.Capability
 
-// Common capabilities
+// Common capabilities.
 const (
-	CapabilityMatrix                             Capability = 0 // Implied by Shader
-	CapabilityShader                             Capability = 1
-	CapabilityFloat16                            Capability = 9    // Required for OpTypeFloat 16
-	CapabilityFloat64                            Capability = 10   // Required for OpTypeFloat 64
-	CapabilityInt64                              Capability = 11   // Required for OpTypeInt 64
-	CapabilityInt16                              Capability = 22   // Required for OpTypeInt 16
-	CapabilityImageGatherExtended                Capability = 25   // Offset on image gather/sample
-	CapabilityInt8                               Capability = 39   // Required for OpTypeInt 8
-	CapabilityLinkage                            Capability = 5    // Import/export linkage
-	CapabilityInt64Atomics                       Capability = 12   // 64-bit atomic operations
-	CapabilityClipDistance                       Capability = 32   // ClipDistance builtin
-	CapabilityImageCubeArray                     Capability = 34   // Cube array storage images
-	CapabilitySampleRateShading                  Capability = 35   // Sample-rate shading
-	CapabilitySampled1D                          Capability = 43   // Sampled 1D images
-	CapabilityImage1D                            Capability = 44   // 1D storage images
-	CapabilitySampledCubeArray                   Capability = 45   // Sampled cube array images
-	CapabilityStorageImageExtendedFormats        Capability = 49   // Extended storage image formats
-	CapabilityImageQuery                         Capability = 50   // Required for OpImageQuerySize/Lod/Levels/Samples
-	CapabilityDerivativeControl                  Capability = 51   // Fine/coarse derivatives
-	CapabilityStorageBuffer16BitAccess           Capability = 4433 // 16-bit storage buffer access
-	CapabilityUniformAndStorageBuffer16BitAccess Capability = 4434 // 16-bit uniform+storage buffer access
-	CapabilityStorageInputOutput16               Capability = 4436 // 16-bit input/output
-	CapabilityMultiView                          Capability = 4439 // MultiView extension
-	CapabilityFragmentBarycentricKHR             Capability = 5284 // Fragment barycentric
-	CapabilityShaderNonUniform                   Capability = 5301 // NonUniform decorations
-	CapabilityAtomicFloat32AddEXT                Capability = 6033 // Float32 atomic add extension
-	CapabilityDotProductInput4x8BitPacked        Capability = 6018 // Required for packed 4x8 dot product
-	CapabilityDotProduct                         Capability = 6019 // Required for integer dot product
-	CapabilityGroupNonUniform                    Capability = 61   // Basic subgroup operations
-	CapabilityGroupNonUniformVote                Capability = 62   // subgroupAll, subgroupAny
-	CapabilityGroupNonUniformArithmetic          Capability = 63   // subgroupAdd, etc.
-	CapabilityGroupNonUniformBallot              Capability = 64   // subgroupBallot
-	CapabilityGroupNonUniformShuffle             Capability = 65   // subgroupShuffle
-	CapabilityGroupNonUniformShuffleRel          Capability = 66   // subgroupShuffleUp/Down
-	CapabilityGroupNonUniformQuad                Capability = 68   // quad operations
-	CapabilityGeometry                           Capability = 2    // Geometry shader
-	CapabilitySubgroupBallotKHR                  Capability = 4423 // KHR subgroup ballot
-	CapabilityInt64ImageEXT                      Capability = 5016 // 64-bit image types (SPV_EXT_shader_image_int64)
+	CapabilityMatrix                             = codegen.CapabilityMatrix
+	CapabilityShader                             = codegen.CapabilityShader
+	CapabilityFloat16                            = codegen.CapabilityFloat16
+	CapabilityFloat64                            = codegen.CapabilityFloat64
+	CapabilityInt64                              = codegen.CapabilityInt64
+	CapabilityInt16                              = codegen.CapabilityInt16
+	CapabilityImageGatherExtended                = codegen.CapabilityImageGatherExtended
+	CapabilityInt8                               = codegen.CapabilityInt8
+	CapabilityLinkage                            = codegen.CapabilityLinkage
+	CapabilityInt64Atomics                       = codegen.CapabilityInt64Atomics
+	CapabilityClipDistance                       = codegen.CapabilityClipDistance
+	CapabilityImageCubeArray                     = codegen.CapabilityImageCubeArray
+	CapabilitySampleRateShading                  = codegen.CapabilitySampleRateShading
+	CapabilitySampled1D                          = codegen.CapabilitySampled1D
+	CapabilityImage1D                            = codegen.CapabilityImage1D
+	CapabilitySampledCubeArray                   = codegen.CapabilitySampledCubeArray
+	CapabilityStorageImageExtendedFormats        = codegen.CapabilityStorageImageExtendedFormats
+	CapabilityImageQuery                         = codegen.CapabilityImageQuery
+	CapabilityDerivativeControl                  = codegen.CapabilityDerivativeControl
+	CapabilityStorageBuffer16BitAccess           = codegen.CapabilityStorageBuffer16BitAccess
+	CapabilityUniformAndStorageBuffer16BitAccess = codegen.CapabilityUniformAndStorageBuffer16BitAccess
+	CapabilityStorageInputOutput16               = codegen.CapabilityStorageInputOutput16
+	CapabilityMultiView                          = codegen.CapabilityMultiView
+	CapabilityFragmentBarycentricKHR             = codegen.CapabilityFragmentBarycentricKHR
+	CapabilityShaderNonUniform                   = codegen.CapabilityShaderNonUniform
+	CapabilityAtomicFloat32AddEXT                = codegen.CapabilityAtomicFloat32AddEXT
+	CapabilityDotProductInput4x8BitPacked        = codegen.CapabilityDotProductInput4x8BitPacked
+	CapabilityDotProduct                         = codegen.CapabilityDotProduct
+	CapabilityGroupNonUniform                    = codegen.CapabilityGroupNonUniform
+	CapabilityGroupNonUniformVote                = codegen.CapabilityGroupNonUniformVote
+	CapabilityGroupNonUniformArithmetic          = codegen.CapabilityGroupNonUniformArithmetic
+	CapabilityGroupNonUniformBallot              = codegen.CapabilityGroupNonUniformBallot
+	CapabilityGroupNonUniformShuffle             = codegen.CapabilityGroupNonUniformShuffle
+	CapabilityGroupNonUniformShuffleRel          = codegen.CapabilityGroupNonUniformShuffleRel
+	CapabilityGroupNonUniformQuad                = codegen.CapabilityGroupNonUniformQuad
+	CapabilityGeometry                           = codegen.CapabilityGeometry
+	CapabilitySubgroupBallotKHR                  = codegen.CapabilitySubgroupBallotKHR
+	CapabilityInt64ImageEXT                      = codegen.CapabilityInt64ImageEXT
 )
 
-// Writer generates SPIR-V from IR.
-type Writer struct {
-	options Options
+// --- Backend (high-level IR → SPIR-V) ---
 
-	// Internal state
-	nextID      uint32
-	typeIDs     map[uint32]uint32
-	constantIDs map[uint32]uint32
+// Backend translates IR to SPIR-V.
+type Backend = codegen.Backend
+
+// NewBackend creates a new SPIR-V backend.
+func NewBackend(options Options) *Backend {
+	return codegen.NewBackend(options)
 }
+
+// --- Low-level builder types (used by examples and advanced consumers) ---
+
+// ModuleBuilder builds complete SPIR-V modules.
+type ModuleBuilder = codegen.ModuleBuilder
+
+// NewModuleBuilder creates a new SPIR-V module builder.
+func NewModuleBuilder(version Version) *ModuleBuilder {
+	return codegen.NewModuleBuilder(version)
+}
+
+// Instruction represents a SPIR-V instruction.
+type Instruction = codegen.Instruction
+
+// InstructionBuilder builds SPIR-V instructions.
+type InstructionBuilder = codegen.InstructionBuilder
+
+// NewInstructionBuilder creates a new instruction builder.
+func NewInstructionBuilder() *InstructionBuilder {
+	return codegen.NewInstructionBuilder()
+}
+
+// Writer generates SPIR-V from IR.
+type Writer = codegen.Writer
 
 // NewWriter creates a new SPIR-V writer.
 func NewWriter(options Options) *Writer {
-	return &Writer{
-		options:     options,
-		nextID:      1,
-		typeIDs:     make(map[uint32]uint32),
-		constantIDs: make(map[uint32]uint32),
-	}
+	return codegen.NewWriter(options)
 }
 
-// SPIR-V magic number and constants
+// --- Block model types ---
+
+// Block represents a SPIR-V basic block under construction.
+type Block = codegen.Block
+
+// NewBlock creates a new block with the given label ID and an empty body.
+func NewBlock(labelID uint32) Block {
+	return codegen.NewBlock(labelID)
+}
+
+// TerminatedBlock is a finalized basic block.
+type TerminatedBlock = codegen.TerminatedBlock
+
+// FunctionBuilder collects terminated blocks for a single SPIR-V function.
+type FunctionBuilder = codegen.FunctionBuilder
+
+// ExpressionEmitter handles expression and statement emission.
+type ExpressionEmitter = codegen.ExpressionEmitter
+
+// BlockExitKind specifies how a block should be terminated.
+type BlockExitKind = codegen.BlockExitKind
+
+// BlockExitKind values.
 const (
-	MagicNumber = 0x07230203
-	GeneratorID = 0x00000000 // Unregistered generator
+	BlockExitReturn  = codegen.BlockExitReturn
+	BlockExitBranch  = codegen.BlockExitBranch
+	BlockExitBreakIf = codegen.BlockExitBreakIf
 )
+
+// BlockExit specifies how a block should end.
+type BlockExit = codegen.BlockExit
+
+// BlockExitDisposition indicates whether writeBlock consumed the provided exit.
+type BlockExitDisposition = codegen.BlockExitDisposition
+
+// BlockExitDisposition values.
+const (
+	ExitUsed      = codegen.ExitUsed
+	ExitDiscarded = codegen.ExitDiscarded
+)
+
+// LoopContext provides break/continue targets for loop bodies.
+type LoopContext = codegen.LoopContext
+
+// --- SPIR-V spec constants and types ---
 
 // OpCode represents a SPIR-V opcode.
-type OpCode uint16
-
-// Common opcodes
-const (
-	OpNop               OpCode = 0
-	OpSource            OpCode = 3
-	OpString            OpCode = 7
-	OpName              OpCode = 5
-	OpMemberName        OpCode = 6
-	OpExtInstImport     OpCode = 11
-	OpMemoryModel       OpCode = 14
-	OpEntryPoint        OpCode = 15
-	OpExecutionMode     OpCode = 16
-	OpCapability        OpCode = 17
-	OpTypeVoid          OpCode = 19
-	OpTypeBool          OpCode = 20
-	OpTypeInt           OpCode = 21
-	OpTypeFloat         OpCode = 22
-	OpTypeVector        OpCode = 23
-	OpTypeMatrix        OpCode = 24
-	OpTypeArray         OpCode = 28
-	OpTypeRuntimeArray  OpCode = 29
-	OpTypeStruct        OpCode = 30
-	OpTypePointer       OpCode = 32
-	OpTypeFunction      OpCode = 33
-	OpConstant          OpCode = 43
-	OpConstantComposite OpCode = 44
-	OpConstantNull      OpCode = 46
-	OpFunction          OpCode = 54
-	OpFunctionParameter OpCode = 55
-	OpFunctionEnd       OpCode = 56
-	OpFunctionCall      OpCode = 57
-	OpVariable          OpCode = 59
-	OpLoad              OpCode = 61
-	OpStore             OpCode = 62
-	OpAccessChain       OpCode = 65
-	OpDecorate          OpCode = 71
-	OpMemberDecorate    OpCode = 72
-	OpLabel             OpCode = 248
-	OpBranch            OpCode = 249
-	OpPhi               OpCode = 245
-	OpReturn            OpCode = 253
-	OpReturnValue       OpCode = 254
-	OpUnreachable       OpCode = 255
-)
+type OpCode = codegen.OpCode
 
 // Decoration represents a SPIR-V decoration.
-type Decoration uint32
-
-// Common decorations
-const (
-	DecorationBlock         Decoration = 2
-	DecorationColMajor      Decoration = 5
-	DecorationRowMajor      Decoration = 4
-	DecorationArrayStride   Decoration = 6
-	DecorationMatrixStride  Decoration = 7
-	DecorationBuiltIn       Decoration = 11
-	DecorationFlat          Decoration = 14
-	DecorationNoPerspective Decoration = 13
-	DecorationCentroid      Decoration = 16
-	DecorationSample        Decoration = 17
-	DecorationNonWritable   Decoration = 24
-	DecorationNonReadable   Decoration = 25
-	DecorationLocation      Decoration = 30
-	DecorationIndex         Decoration = 32 // For dual-source blending
-	DecorationBinding       Decoration = 33
-	DecorationDescriptorSet Decoration = 34
-	DecorationOffset        Decoration = 35
-	DecorationNonUniform    Decoration = 5300 // NonUniformEXT
-)
+type Decoration = codegen.Decoration
 
 // BuiltIn represents a SPIR-V built-in decoration value.
-type BuiltIn uint32
-
-// SPIR-V built-in values (used with DecorationBuiltIn).
-const (
-	BuiltInPosition             BuiltIn = 0
-	BuiltInPointSize            BuiltIn = 1
-	BuiltInClipDistance         BuiltIn = 3
-	BuiltInCullDistance         BuiltIn = 4
-	BuiltInVertexID             BuiltIn = 5
-	BuiltInInstanceID           BuiltIn = 6
-	BuiltInPrimitiveID          BuiltIn = 7
-	BuiltInInvocationID         BuiltIn = 8
-	BuiltInLayer                BuiltIn = 9
-	BuiltInViewportIndex        BuiltIn = 10
-	BuiltInTessLevelOuter       BuiltIn = 11
-	BuiltInTessLevelInner       BuiltIn = 12
-	BuiltInTessCoord            BuiltIn = 13
-	BuiltInPatchVertices        BuiltIn = 14
-	BuiltInFragCoord            BuiltIn = 15
-	BuiltInPointCoord           BuiltIn = 16
-	BuiltInFrontFacing          BuiltIn = 17
-	BuiltInSampleID             BuiltIn = 18
-	BuiltInSamplePosition       BuiltIn = 19
-	BuiltInSampleMask           BuiltIn = 20
-	BuiltInFragDepth            BuiltIn = 22
-	BuiltInHelperInvocation     BuiltIn = 23
-	BuiltInNumWorkgroups        BuiltIn = 24
-	BuiltInWorkgroupSize        BuiltIn = 25
-	BuiltInWorkgroupID          BuiltIn = 26
-	BuiltInLocalInvocationID    BuiltIn = 27
-	BuiltInGlobalInvocationID   BuiltIn = 28
-	BuiltInLocalInvocationIndex BuiltIn = 29
-	BuiltInVertexIndex          BuiltIn = 42
-	BuiltInInstanceIndex        BuiltIn = 43
-	BuiltInSubgroupSize         BuiltIn = 36
-	BuiltInSubgroupLocalInvID   BuiltIn = 41
-	BuiltInNumSubgroups         BuiltIn = 38
-	BuiltInSubgroupID           BuiltIn = 40
-	BuiltInViewIndex            BuiltIn = 4440
-	BuiltInBaryCoordKHR         BuiltIn = 5286
-)
+type BuiltIn = codegen.BuiltIn
 
 // ExecutionModel represents a SPIR-V execution model.
-type ExecutionModel uint32
-
-// Common execution models
-const (
-	ExecutionModelVertex                 ExecutionModel = 0
-	ExecutionModelTessellationControl    ExecutionModel = 1
-	ExecutionModelTessellationEvaluation ExecutionModel = 2
-	ExecutionModelGeometry               ExecutionModel = 3
-	ExecutionModelFragment               ExecutionModel = 4
-	ExecutionModelGLCompute              ExecutionModel = 5
-	ExecutionModelKernel                 ExecutionModel = 6
-)
+type ExecutionModel = codegen.ExecutionModel
 
 // ExecutionMode represents a SPIR-V execution mode.
-type ExecutionMode uint32
-
-// Common execution modes
-const (
-	ExecutionModeInvocations              ExecutionMode = 0
-	ExecutionModeSpacingEqual             ExecutionMode = 1
-	ExecutionModeSpacingFractionalEven    ExecutionMode = 2
-	ExecutionModeSpacingFractionalOdd     ExecutionMode = 3
-	ExecutionModeVertexOrderCw            ExecutionMode = 4
-	ExecutionModeVertexOrderCcw           ExecutionMode = 5
-	ExecutionModePixelCenterInteger       ExecutionMode = 6
-	ExecutionModeOriginUpperLeft          ExecutionMode = 7
-	ExecutionModeOriginLowerLeft          ExecutionMode = 8
-	ExecutionModeEarlyFragmentTests       ExecutionMode = 9
-	ExecutionModePointMode                ExecutionMode = 10
-	ExecutionModeXfb                      ExecutionMode = 11
-	ExecutionModeDepthReplacing           ExecutionMode = 12
-	ExecutionModeDepthGreater             ExecutionMode = 14
-	ExecutionModeDepthLess                ExecutionMode = 15
-	ExecutionModeDepthUnchanged           ExecutionMode = 16
-	ExecutionModeLocalSize                ExecutionMode = 17
-	ExecutionModeLocalSizeHint            ExecutionMode = 18
-	ExecutionModeInputPoints              ExecutionMode = 19
-	ExecutionModeInputLines               ExecutionMode = 20
-	ExecutionModeInputLinesAdjacency      ExecutionMode = 21
-	ExecutionModeTriangles                ExecutionMode = 22
-	ExecutionModeInputTrianglesAdjacency  ExecutionMode = 23
-	ExecutionModeQuads                    ExecutionMode = 24
-	ExecutionModeIsolines                 ExecutionMode = 25
-	ExecutionModeOutputVertices           ExecutionMode = 26
-	ExecutionModeOutputPoints             ExecutionMode = 27
-	ExecutionModeOutputLineStrip          ExecutionMode = 28
-	ExecutionModeOutputTriangleStrip      ExecutionMode = 29
-	ExecutionModeVecTypeHint              ExecutionMode = 30
-	ExecutionModeContractionOff           ExecutionMode = 31
-	ExecutionModeInitializer              ExecutionMode = 33
-	ExecutionModeFinalizer                ExecutionMode = 34
-	ExecutionModeSubgroupSize             ExecutionMode = 35
-	ExecutionModeSubgroupsPerWorkgroup    ExecutionMode = 36
-	ExecutionModeSubgroupsPerWorkgroupID  ExecutionMode = 37
-	ExecutionModeLocalSizeID              ExecutionMode = 38
-	ExecutionModeLocalSizeHintID          ExecutionMode = 39
-	ExecutionModePostDepthCoverage        ExecutionMode = 4446
-	ExecutionModeDenormPreserve           ExecutionMode = 4459
-	ExecutionModeDenormFlushToZero        ExecutionMode = 4460
-	ExecutionModeSignedZeroInfNanPreserve ExecutionMode = 4461
-	ExecutionModeRoundingModeRTE          ExecutionMode = 4462
-	ExecutionModeRoundingModeRTZ          ExecutionMode = 4463
-)
+type ExecutionMode = codegen.ExecutionMode
 
 // StorageClass represents a SPIR-V storage class.
-type StorageClass uint32
-
-// Common storage classes
-const (
-	StorageClassUniformConstant         StorageClass = 0
-	StorageClassInput                   StorageClass = 1
-	StorageClassUniform                 StorageClass = 2
-	StorageClassOutput                  StorageClass = 3
-	StorageClassWorkgroup               StorageClass = 4
-	StorageClassCrossWorkgroup          StorageClass = 5
-	StorageClassPrivate                 StorageClass = 6
-	StorageClassFunction                StorageClass = 7
-	StorageClassGeneric                 StorageClass = 8
-	StorageClassPushConstant            StorageClass = 9
-	StorageClassAtomicCounter           StorageClass = 10
-	StorageClassImage                   StorageClass = 11
-	StorageClassStorageBuffer           StorageClass = 12
-	StorageClassTaskPayloadWorkgroupEXT StorageClass = 5328
-)
+type StorageClass = codegen.StorageClass
 
 // AddressingModel represents a SPIR-V addressing model.
-type AddressingModel uint32
-
-// Common addressing models
-const (
-	AddressingModelLogical    AddressingModel = 0
-	AddressingModelPhysical32 AddressingModel = 1
-	AddressingModelPhysical64 AddressingModel = 2
-)
+type AddressingModel = codegen.AddressingModel
 
 // MemoryModel represents a SPIR-V memory model.
-type MemoryModel uint32
-
-// Common memory models
-const (
-	MemoryModelSimple  MemoryModel = 0
-	MemoryModelGLSL450 MemoryModel = 1
-	MemoryModelOpenCL  MemoryModel = 2
-	MemoryModelVulkan  MemoryModel = 3
-)
+type MemoryModel = codegen.MemoryModel
 
 // FunctionControl represents a SPIR-V function control.
-type FunctionControl uint32
+type FunctionControl = codegen.FunctionControl
 
-// Common function control flags
+// SelectionControl flags for OpSelectionMerge.
+type SelectionControl = codegen.SelectionControl
+
+// LoopControl flags for OpLoopMerge.
+type LoopControl = codegen.LoopControl
+
+// ImageFormat represents a SPIR-V image format.
+type ImageFormat = codegen.ImageFormat
+
+// SPIR-V magic number and constants.
 const (
-	FunctionControlNone       FunctionControl = 0x0
-	FunctionControlInline     FunctionControl = 0x1
-	FunctionControlDontInline FunctionControl = 0x2
-	FunctionControlPure       FunctionControl = 0x4
-	FunctionControlConst      FunctionControl = 0x8
+	MagicNumber = codegen.MagicNumber
+	GeneratorID = codegen.GeneratorID
 )
 
-// OpExtension represents OpExtension opcode.
-const OpExtension OpCode = 10
+// --- Opcodes ---
 
-// Arithmetic opcodes
+// Common opcodes.
 const (
-	OpFNegate           OpCode = 127 // Float negation
-	OpSNegate           OpCode = 126 // Signed integer negation
-	OpFAdd              OpCode = 129 // Float addition
-	OpFSub              OpCode = 131 // Float subtraction
-	OpFMul              OpCode = 133 // Float multiplication
-	OpUDiv              OpCode = 134 // Unsigned integer division
-	OpSDiv              OpCode = 135 // Signed integer division
-	OpFDiv              OpCode = 136 // Float division
-	OpUMod              OpCode = 137 // Unsigned integer modulo
-	OpSRem              OpCode = 138 // Signed integer remainder
-	OpSMod              OpCode = 139 // Signed integer modulo
-	OpFRem              OpCode = 140 // Float remainder
-	OpFMod              OpCode = 141 // Float modulo
-	OpVectorTimesScalar OpCode = 142 // Vector-scalar float multiplication
-	OpMatrixTimesScalar OpCode = 143 // Matrix-scalar float multiplication
-	OpVectorTimesMatrix OpCode = 144 // Vector-matrix multiplication
-	OpMatrixTimesVector OpCode = 145 // Matrix-vector multiplication
-	OpMatrixTimesMatrix OpCode = 146 // Matrix-matrix multiplication
-	OpIAdd              OpCode = 128 // Integer addition
-	OpISub              OpCode = 130 // Integer subtraction
-	OpIMul              OpCode = 132 // Integer multiplication
+	OpNop               = codegen.OpNop
+	OpSource            = codegen.OpSource
+	OpString            = codegen.OpString
+	OpName              = codegen.OpName
+	OpMemberName        = codegen.OpMemberName
+	OpExtInstImport     = codegen.OpExtInstImport
+	OpMemoryModel       = codegen.OpMemoryModel
+	OpEntryPoint        = codegen.OpEntryPoint
+	OpExecutionMode     = codegen.OpExecutionMode
+	OpCapability        = codegen.OpCapability
+	OpTypeVoid          = codegen.OpTypeVoid
+	OpTypeBool          = codegen.OpTypeBool
+	OpTypeInt           = codegen.OpTypeInt
+	OpTypeFloat         = codegen.OpTypeFloat
+	OpTypeVector        = codegen.OpTypeVector
+	OpTypeMatrix        = codegen.OpTypeMatrix
+	OpTypeArray         = codegen.OpTypeArray
+	OpTypeRuntimeArray  = codegen.OpTypeRuntimeArray
+	OpTypeStruct        = codegen.OpTypeStruct
+	OpTypePointer       = codegen.OpTypePointer
+	OpTypeFunction      = codegen.OpTypeFunction
+	OpConstant          = codegen.OpConstant
+	OpConstantComposite = codegen.OpConstantComposite
+	OpConstantNull      = codegen.OpConstantNull
+	OpFunction          = codegen.OpFunction
+	OpFunctionParameter = codegen.OpFunctionParameter
+	OpFunctionEnd       = codegen.OpFunctionEnd
+	OpFunctionCall      = codegen.OpFunctionCall
+	OpVariable          = codegen.OpVariable
+	OpLoad              = codegen.OpLoad
+	OpStore             = codegen.OpStore
+	OpAccessChain       = codegen.OpAccessChain
+	OpDecorate          = codegen.OpDecorate
+	OpMemberDecorate    = codegen.OpMemberDecorate
+	OpLabel             = codegen.OpLabel
+	OpBranch            = codegen.OpBranch
+	OpPhi               = codegen.OpPhi
+	OpReturn            = codegen.OpReturn
+	OpReturnValue       = codegen.OpReturnValue
+	OpUnreachable       = codegen.OpUnreachable
+	OpExtension         = codegen.OpExtension
 )
 
-// Comparison opcodes
+// Arithmetic opcodes.
 const (
-	OpFOrdEqual            OpCode = 180 // Float ordered equal
-	OpFOrdNotEqual         OpCode = 182 // Float ordered not equal
-	OpFOrdLessThan         OpCode = 184 // Float ordered less than
-	OpFOrdGreaterThan      OpCode = 186 // Float ordered greater than
-	OpFOrdLessThanEqual    OpCode = 188 // Float ordered less than or equal
-	OpFOrdGreaterThanEqual OpCode = 190 // Float ordered greater than or equal
-	OpIEqual               OpCode = 170 // Integer equal
-	OpINotEqual            OpCode = 171 // Integer not equal
-	OpSLessThan            OpCode = 177 // Signed integer less than
-	OpSLessThanEqual       OpCode = 179 // Signed integer less than or equal
-	OpSGreaterThan         OpCode = 173 // Signed integer greater than
-	OpSGreaterThanEqual    OpCode = 175 // Signed integer greater than or equal
-	OpULessThan            OpCode = 176 // Unsigned integer less than
-	OpULessThanEqual       OpCode = 178 // Unsigned integer less than or equal
-	OpUGreaterThan         OpCode = 172 // Unsigned integer greater than
-	OpUGreaterThanEqual    OpCode = 174 // Unsigned integer greater than or equal
+	OpFNegate           = codegen.OpFNegate
+	OpSNegate           = codegen.OpSNegate
+	OpFAdd              = codegen.OpFAdd
+	OpFSub              = codegen.OpFSub
+	OpFMul              = codegen.OpFMul
+	OpUDiv              = codegen.OpUDiv
+	OpSDiv              = codegen.OpSDiv
+	OpFDiv              = codegen.OpFDiv
+	OpUMod              = codegen.OpUMod
+	OpSRem              = codegen.OpSRem
+	OpSMod              = codegen.OpSMod
+	OpFRem              = codegen.OpFRem
+	OpFMod              = codegen.OpFMod
+	OpVectorTimesScalar = codegen.OpVectorTimesScalar
+	OpMatrixTimesScalar = codegen.OpMatrixTimesScalar
+	OpVectorTimesMatrix = codegen.OpVectorTimesMatrix
+	OpMatrixTimesVector = codegen.OpMatrixTimesVector
+	OpMatrixTimesMatrix = codegen.OpMatrixTimesMatrix
+	OpIAdd              = codegen.OpIAdd
+	OpISub              = codegen.OpISub
+	OpIMul              = codegen.OpIMul
 )
 
-// Logical opcodes
+// Comparison opcodes.
 const (
-	OpLogicalEqual    OpCode = 164 // Logical equal
-	OpLogicalNotEqual OpCode = 165 // Logical not equal
-	OpLogicalOr       OpCode = 166 // Logical or
-	OpLogicalAnd      OpCode = 167 // Logical and
-	OpLogicalNot      OpCode = 168 // Logical not
-	OpSelect          OpCode = 169 // Select between values
-	OpNot             OpCode = 200 // Bitwise not
-	OpAny             OpCode = 154 // True if any component of boolean vector is true
-	OpAll             OpCode = 155 // True if all components of boolean vector are true
-	OpIsNan           OpCode = 156 // True if value is NaN
-	OpIsInf           OpCode = 157 // True if value is Inf
+	OpFOrdEqual            = codegen.OpFOrdEqual
+	OpFOrdNotEqual         = codegen.OpFOrdNotEqual
+	OpFOrdLessThan         = codegen.OpFOrdLessThan
+	OpFOrdGreaterThan      = codegen.OpFOrdGreaterThan
+	OpFOrdLessThanEqual    = codegen.OpFOrdLessThanEqual
+	OpFOrdGreaterThanEqual = codegen.OpFOrdGreaterThanEqual
+	OpIEqual               = codegen.OpIEqual
+	OpINotEqual            = codegen.OpINotEqual
+	OpSLessThan            = codegen.OpSLessThan
+	OpSLessThanEqual       = codegen.OpSLessThanEqual
+	OpSGreaterThan         = codegen.OpSGreaterThan
+	OpSGreaterThanEqual    = codegen.OpSGreaterThanEqual
+	OpULessThan            = codegen.OpULessThan
+	OpULessThanEqual       = codegen.OpULessThanEqual
+	OpUGreaterThan         = codegen.OpUGreaterThan
+	OpUGreaterThanEqual    = codegen.OpUGreaterThanEqual
 )
 
-// Composite opcodes
+// Logical opcodes.
 const (
-	OpVectorExtractDynamic OpCode = 77 // Extract from vector with dynamic index
-	OpVectorShuffle        OpCode = 79 // Shuffle vector components
-	OpCompositeConstruct   OpCode = 80 // Construct composite
-	OpCompositeExtract     OpCode = 81 // Extract from composite
+	OpLogicalEqual    = codegen.OpLogicalEqual
+	OpLogicalNotEqual = codegen.OpLogicalNotEqual
+	OpLogicalOr       = codegen.OpLogicalOr
+	OpLogicalAnd      = codegen.OpLogicalAnd
+	OpLogicalNot      = codegen.OpLogicalNot
+	OpSelect          = codegen.OpSelect
+	OpNot             = codegen.OpNot
+	OpAny             = codegen.OpAny
+	OpAll             = codegen.OpAll
+	OpIsNan           = codegen.OpIsNan
+	OpIsInf           = codegen.OpIsInf
 )
 
-// Bitwise opcodes
+// Composite opcodes.
 const (
-	OpShiftRightLogical    OpCode = 194 // Shift right logical
-	OpShiftRightArithmetic OpCode = 195 // Shift right arithmetic
-	OpShiftLeftLogical     OpCode = 196 // Shift left logical
-	OpBitwiseOr            OpCode = 197 // Bitwise OR
-	OpBitwiseXor           OpCode = 198 // Bitwise XOR
-	OpBitwiseAnd           OpCode = 199 // Bitwise AND
-	OpBitFieldInsert       OpCode = 201 // Insert bit field
-	OpBitFieldSExtract     OpCode = 202 // Extract signed bit field
-	OpBitFieldUExtract     OpCode = 203 // Extract unsigned bit field
-	OpBitReverse           OpCode = 204 // Reverse bits
-	OpBitCount             OpCode = 205 // Count set bits
+	OpVectorExtractDynamic = codegen.OpVectorExtractDynamic
+	OpVectorShuffle        = codegen.OpVectorShuffle
+	OpCompositeConstruct   = codegen.OpCompositeConstruct
+	OpCompositeExtract     = codegen.OpCompositeExtract
 )
 
-// Control flow opcodes
+// Bitwise opcodes.
 const (
-	OpSelectionMerge    OpCode = 247 // Selection merge point
-	OpLoopMerge         OpCode = 246 // Loop merge point
-	OpBranchConditional OpCode = 250 // Conditional branch
-	OpSwitch            OpCode = 251 // Switch statement
-	OpKill              OpCode = 252 // Fragment discard
+	OpShiftRightLogical    = codegen.OpShiftRightLogical
+	OpShiftRightArithmetic = codegen.OpShiftRightArithmetic
+	OpShiftLeftLogical     = codegen.OpShiftLeftLogical
+	OpBitwiseOr            = codegen.OpBitwiseOr
+	OpBitwiseXor           = codegen.OpBitwiseXor
+	OpBitwiseAnd           = codegen.OpBitwiseAnd
+	OpBitFieldInsert       = codegen.OpBitFieldInsert
+	OpBitFieldSExtract     = codegen.OpBitFieldSExtract
+	OpBitFieldUExtract     = codegen.OpBitFieldUExtract
+	OpBitReverse           = codegen.OpBitReverse
+	OpBitCount             = codegen.OpBitCount
 )
 
-// Derivative opcodes
+// Control flow opcodes.
 const (
-	OpDPdx         OpCode = 207 // Derivative in X
-	OpDPdy         OpCode = 208 // Derivative in Y
-	OpFwidth       OpCode = 209 // Sum of absolute derivatives
-	OpDPdxFine     OpCode = 210 // Fine derivative in X
-	OpDPdyFine     OpCode = 211 // Fine derivative in Y
-	OpFwidthFine   OpCode = 212 // Fine fwidth
-	OpDPdxCoarse   OpCode = 213 // Coarse derivative in X
-	OpDPdyCoarse   OpCode = 214 // Coarse derivative in Y
-	OpFwidthCoarse OpCode = 215 // Coarse fwidth
+	OpSelectionMerge    = codegen.OpSelectionMerge
+	OpLoopMerge         = codegen.OpLoopMerge
+	OpBranchConditional = codegen.OpBranchConditional
+	OpSwitch            = codegen.OpSwitch
+	OpKill              = codegen.OpKill
 )
 
-// Conversion opcodes
+// Derivative opcodes.
 const (
-	OpConvertFToU   OpCode = 109 // Float to unsigned int
-	OpConvertFToS   OpCode = 110 // Float to signed int
-	OpConvertSToF   OpCode = 111 // Signed int to float
-	OpConvertUToF   OpCode = 112 // Unsigned int to float
-	OpUConvert      OpCode = 113 // Unsigned int width conversion
-	OpSConvert      OpCode = 114 // Signed int width conversion
-	OpFConvert      OpCode = 115 // Float width conversion
-	OpQuantizeToF16 OpCode = 116 // Quantize float to f16 precision
-	OpBitcast       OpCode = 124 // Bitcast between types of same width
+	OpDPdx         = codegen.OpDPdx
+	OpDPdy         = codegen.OpDPdy
+	OpFwidth       = codegen.OpFwidth
+	OpDPdxFine     = codegen.OpDPdxFine
+	OpDPdyFine     = codegen.OpDPdyFine
+	OpFwidthFine   = codegen.OpFwidthFine
+	OpDPdxCoarse   = codegen.OpDPdxCoarse
+	OpDPdyCoarse   = codegen.OpDPdyCoarse
+	OpFwidthCoarse = codegen.OpFwidthCoarse
 )
 
-// Extended instruction set opcodes
+// Conversion opcodes.
 const (
-	OpExtInst OpCode = 12 // Extended instruction
+	OpConvertFToU   = codegen.OpConvertFToU
+	OpConvertFToS   = codegen.OpConvertFToS
+	OpConvertSToF   = codegen.OpConvertSToF
+	OpConvertUToF   = codegen.OpConvertUToF
+	OpUConvert      = codegen.OpUConvert
+	OpSConvert      = codegen.OpSConvert
+	OpFConvert      = codegen.OpFConvert
+	OpQuantizeToF16 = codegen.OpQuantizeToF16
+	OpBitcast       = codegen.OpBitcast
 )
 
-// Atomic operation opcodes
+// Extended instruction set opcodes.
 const (
-	OpAtomicLoad        OpCode = 227  // Atomic load
-	OpAtomicStore       OpCode = 228  // Atomic store
-	OpAtomicExchange    OpCode = 229  // Atomic exchange
-	OpAtomicCompareExch OpCode = 230  // Atomic compare-exchange
-	OpAtomicIIncrement  OpCode = 232  // Atomic integer increment
-	OpAtomicIDecrement  OpCode = 233  // Atomic integer decrement
-	OpAtomicIAdd        OpCode = 234  // Atomic integer add
-	OpAtomicFAddEXT     OpCode = 6035 // Atomic float add (SPV_EXT_shader_atomic_float_add)
-	OpAtomicISub        OpCode = 235  // Atomic integer subtract
-	OpAtomicSMin        OpCode = 236  // Atomic signed min
-	OpAtomicUMin        OpCode = 237  // Atomic unsigned min
-	OpAtomicSMax        OpCode = 238  // Atomic signed max
-	OpAtomicUMax        OpCode = 239  // Atomic unsigned max
-	OpAtomicAnd         OpCode = 240  // Atomic bitwise and
-	OpAtomicOr          OpCode = 241  // Atomic bitwise or
-	OpAtomicXor         OpCode = 242  // Atomic bitwise xor
+	OpExtInst = codegen.OpExtInst
 )
 
-// Integer dot product opcodes (SPV_KHR_integer_dot_product)
+// Atomic operation opcodes.
 const (
-	OpSDotKHR OpCode = 4450 // Signed integer dot product
-	OpUDotKHR OpCode = 4451 // Unsigned integer dot product
-
-	// PackedVectorFormat4x8Bit is the packed vector format for 4x8 bit integers.
-	PackedVectorFormat4x8Bit uint32 = 0
+	OpAtomicLoad        = codegen.OpAtomicLoad
+	OpAtomicStore       = codegen.OpAtomicStore
+	OpAtomicExchange    = codegen.OpAtomicExchange
+	OpAtomicCompareExch = codegen.OpAtomicCompareExch
+	OpAtomicIIncrement  = codegen.OpAtomicIIncrement
+	OpAtomicIDecrement  = codegen.OpAtomicIDecrement
+	OpAtomicIAdd        = codegen.OpAtomicIAdd
+	OpAtomicFAddEXT     = codegen.OpAtomicFAddEXT
+	OpAtomicISub        = codegen.OpAtomicISub
+	OpAtomicSMin        = codegen.OpAtomicSMin
+	OpAtomicUMin        = codegen.OpAtomicUMin
+	OpAtomicSMax        = codegen.OpAtomicSMax
+	OpAtomicUMax        = codegen.OpAtomicUMax
+	OpAtomicAnd         = codegen.OpAtomicAnd
+	OpAtomicOr          = codegen.OpAtomicOr
+	OpAtomicXor         = codegen.OpAtomicXor
 )
 
-// Memory scope for atomic operations
+// Integer dot product opcodes.
 const (
-	ScopeDevice    uint32 = 1 // Visible to all invocations in the device
-	ScopeWorkgroup uint32 = 2 // Visible to all invocations in the workgroup
-	ScopeSubgroup  uint32 = 3 // Visible to all invocations in the subgroup
+	OpSDotKHR = codegen.OpSDotKHR
+	OpUDotKHR = codegen.OpUDotKHR
+
+	PackedVectorFormat4x8Bit = codegen.PackedVectorFormat4x8Bit
 )
 
-// Memory semantics for atomic operations
+// Memory scope for atomic operations.
 const (
-	MemorySemanticsNone                uint32 = 0x0
-	MemorySemanticsAcquire             uint32 = 0x2
-	MemorySemanticsRelease             uint32 = 0x4
-	MemorySemanticsAcquireRelease      uint32 = 0x8
-	MemorySemanticsUniformMemory       uint32 = 0x40
-	MemorySemanticsSubgroupMemory      uint32 = 0x80
-	MemorySemanticsWorkgroupMemory     uint32 = 0x100
-	MemorySemanticsAtomicCounterMemory uint32 = 0x400
-	MemorySemanticsImageMemory         uint32 = 0x800
+	ScopeDevice    = codegen.ScopeDevice
+	ScopeWorkgroup = codegen.ScopeWorkgroup
+	ScopeSubgroup  = codegen.ScopeSubgroup
 )
 
-// Barrier opcodes
+// Memory semantics for atomic operations.
 const (
-	OpControlBarrier OpCode = 224 // Control barrier (execution + memory)
-	OpMemoryBarrier  OpCode = 225 // Memory barrier only
+	MemorySemanticsNone                = codegen.MemorySemanticsNone
+	MemorySemanticsAcquire             = codegen.MemorySemanticsAcquire
+	MemorySemanticsRelease             = codegen.MemorySemanticsRelease
+	MemorySemanticsAcquireRelease      = codegen.MemorySemanticsAcquireRelease
+	MemorySemanticsUniformMemory       = codegen.MemorySemanticsUniformMemory
+	MemorySemanticsSubgroupMemory      = codegen.MemorySemanticsSubgroupMemory
+	MemorySemanticsWorkgroupMemory     = codegen.MemorySemanticsWorkgroupMemory
+	MemorySemanticsAtomicCounterMemory = codegen.MemorySemanticsAtomicCounterMemory
+	MemorySemanticsImageMemory         = codegen.MemorySemanticsImageMemory
 )
 
-// Subgroup opcodes (SPV_KHR_subgroup_vote, SPV_KHR_subgroup_ballot, etc.)
+// Barrier opcodes.
 const (
-	OpGroupNonUniformElect            OpCode = 333
-	OpGroupNonUniformAll              OpCode = 334
-	OpGroupNonUniformAny              OpCode = 335
-	OpGroupNonUniformAllEqual         OpCode = 336
-	OpGroupNonUniformBroadcast        OpCode = 337
-	OpGroupNonUniformBroadcastFirst   OpCode = 338
-	OpGroupNonUniformBallot           OpCode = 339
-	OpGroupNonUniformInverseBallot    OpCode = 340
-	OpGroupNonUniformBallotBitExtract OpCode = 341
-	OpGroupNonUniformBallotBitCount   OpCode = 342
-	OpGroupNonUniformBallotFindLSB    OpCode = 343
-	OpGroupNonUniformBallotFindMSB    OpCode = 344
-	OpGroupNonUniformShuffle          OpCode = 345
-	OpGroupNonUniformShuffleXor       OpCode = 346
-	OpGroupNonUniformShuffleUp        OpCode = 347
-	OpGroupNonUniformShuffleDown      OpCode = 348
-	OpGroupNonUniformIAdd             OpCode = 349
-	OpGroupNonUniformFAdd             OpCode = 350
-	OpGroupNonUniformIMul             OpCode = 351
-	OpGroupNonUniformFMul             OpCode = 352
-	OpGroupNonUniformSMin             OpCode = 353
-	OpGroupNonUniformUMin             OpCode = 354
-	OpGroupNonUniformFMin             OpCode = 355
-	OpGroupNonUniformSMax             OpCode = 356
-	OpGroupNonUniformUMax             OpCode = 357
-	OpGroupNonUniformFMax             OpCode = 358
-	OpGroupNonUniformBitwiseAnd       OpCode = 359
-	OpGroupNonUniformBitwiseOr        OpCode = 360
-	OpGroupNonUniformBitwiseXor       OpCode = 361
-	OpGroupNonUniformLogicalAnd       OpCode = 362
-	OpGroupNonUniformLogicalOr        OpCode = 363
-	OpGroupNonUniformLogicalXor       OpCode = 364
-	OpGroupNonUniformQuadBroadcast    OpCode = 365
-	OpGroupNonUniformQuadSwap         OpCode = 366
+	OpControlBarrier = codegen.OpControlBarrier
+	OpMemoryBarrier  = codegen.OpMemoryBarrier
+)
+
+// Subgroup opcodes.
+const (
+	OpGroupNonUniformElect            = codegen.OpGroupNonUniformElect
+	OpGroupNonUniformAll              = codegen.OpGroupNonUniformAll
+	OpGroupNonUniformAny              = codegen.OpGroupNonUniformAny
+	OpGroupNonUniformAllEqual         = codegen.OpGroupNonUniformAllEqual
+	OpGroupNonUniformBroadcast        = codegen.OpGroupNonUniformBroadcast
+	OpGroupNonUniformBroadcastFirst   = codegen.OpGroupNonUniformBroadcastFirst
+	OpGroupNonUniformBallot           = codegen.OpGroupNonUniformBallot
+	OpGroupNonUniformInverseBallot    = codegen.OpGroupNonUniformInverseBallot
+	OpGroupNonUniformBallotBitExtract = codegen.OpGroupNonUniformBallotBitExtract
+	OpGroupNonUniformBallotBitCount   = codegen.OpGroupNonUniformBallotBitCount
+	OpGroupNonUniformBallotFindLSB    = codegen.OpGroupNonUniformBallotFindLSB
+	OpGroupNonUniformBallotFindMSB    = codegen.OpGroupNonUniformBallotFindMSB
+	OpGroupNonUniformShuffle          = codegen.OpGroupNonUniformShuffle
+	OpGroupNonUniformShuffleXor       = codegen.OpGroupNonUniformShuffleXor
+	OpGroupNonUniformShuffleUp        = codegen.OpGroupNonUniformShuffleUp
+	OpGroupNonUniformShuffleDown      = codegen.OpGroupNonUniformShuffleDown
+	OpGroupNonUniformIAdd             = codegen.OpGroupNonUniformIAdd
+	OpGroupNonUniformFAdd             = codegen.OpGroupNonUniformFAdd
+	OpGroupNonUniformIMul             = codegen.OpGroupNonUniformIMul
+	OpGroupNonUniformFMul             = codegen.OpGroupNonUniformFMul
+	OpGroupNonUniformSMin             = codegen.OpGroupNonUniformSMin
+	OpGroupNonUniformUMin             = codegen.OpGroupNonUniformUMin
+	OpGroupNonUniformFMin             = codegen.OpGroupNonUniformFMin
+	OpGroupNonUniformSMax             = codegen.OpGroupNonUniformSMax
+	OpGroupNonUniformUMax             = codegen.OpGroupNonUniformUMax
+	OpGroupNonUniformFMax             = codegen.OpGroupNonUniformFMax
+	OpGroupNonUniformBitwiseAnd       = codegen.OpGroupNonUniformBitwiseAnd
+	OpGroupNonUniformBitwiseOr        = codegen.OpGroupNonUniformBitwiseOr
+	OpGroupNonUniformBitwiseXor       = codegen.OpGroupNonUniformBitwiseXor
+	OpGroupNonUniformLogicalAnd       = codegen.OpGroupNonUniformLogicalAnd
+	OpGroupNonUniformLogicalOr        = codegen.OpGroupNonUniformLogicalOr
+	OpGroupNonUniformLogicalXor       = codegen.OpGroupNonUniformLogicalXor
+	OpGroupNonUniformQuadBroadcast    = codegen.OpGroupNonUniformQuadBroadcast
+	OpGroupNonUniformQuadSwap         = codegen.OpGroupNonUniformQuadSwap
 )
 
 // GroupOperation for subgroup collective operations.
 const (
-	GroupOperationReduce        uint32 = 0
-	GroupOperationInclusiveScan uint32 = 1
-	GroupOperationExclusiveScan uint32 = 2
+	GroupOperationReduce        = codegen.GroupOperationReduce
+	GroupOperationInclusiveScan = codegen.GroupOperationInclusiveScan
+	GroupOperationExclusiveScan = codegen.GroupOperationExclusiveScan
 )
 
-// SelectionControl flags for OpSelectionMerge
-type SelectionControl uint32
+// --- Decorations ---
 
 const (
-	SelectionControlNone        SelectionControl = 0x0
-	SelectionControlFlatten     SelectionControl = 0x1
-	SelectionControlDontFlatten SelectionControl = 0x2
+	DecorationBlock         = codegen.DecorationBlock
+	DecorationColMajor      = codegen.DecorationColMajor
+	DecorationRowMajor      = codegen.DecorationRowMajor
+	DecorationArrayStride   = codegen.DecorationArrayStride
+	DecorationMatrixStride  = codegen.DecorationMatrixStride
+	DecorationBuiltIn       = codegen.DecorationBuiltIn
+	DecorationFlat          = codegen.DecorationFlat
+	DecorationNoPerspective = codegen.DecorationNoPerspective
+	DecorationCentroid      = codegen.DecorationCentroid
+	DecorationSample        = codegen.DecorationSample
+	DecorationNonWritable   = codegen.DecorationNonWritable
+	DecorationNonReadable   = codegen.DecorationNonReadable
+	DecorationLocation      = codegen.DecorationLocation
+	DecorationIndex         = codegen.DecorationIndex
+	DecorationBinding       = codegen.DecorationBinding
+	DecorationDescriptorSet = codegen.DecorationDescriptorSet
+	DecorationOffset        = codegen.DecorationOffset
+	DecorationNonUniform    = codegen.DecorationNonUniform
 )
 
-// LoopControl flags for OpLoopMerge
-type LoopControl uint32
+// --- BuiltIn values ---
 
 const (
-	LoopControlNone               LoopControl = 0x0
-	LoopControlUnroll             LoopControl = 0x1
-	LoopControlDontUnroll         LoopControl = 0x2
-	LoopControlDependencyInfinite LoopControl = 0x4
-	LoopControlDependencyLength   LoopControl = 0x8
-	LoopControlMinIterations      LoopControl = 0x10
-	LoopControlMaxIterations      LoopControl = 0x20
-	LoopControlIterationMultiple  LoopControl = 0x40
-	LoopControlPeelCount          LoopControl = 0x80
-	LoopControlPartialCount       LoopControl = 0x100
+	BuiltInPosition             = codegen.BuiltInPosition
+	BuiltInPointSize            = codegen.BuiltInPointSize
+	BuiltInClipDistance         = codegen.BuiltInClipDistance
+	BuiltInCullDistance         = codegen.BuiltInCullDistance
+	BuiltInVertexID             = codegen.BuiltInVertexID
+	BuiltInInstanceID           = codegen.BuiltInInstanceID
+	BuiltInPrimitiveID          = codegen.BuiltInPrimitiveID
+	BuiltInInvocationID         = codegen.BuiltInInvocationID
+	BuiltInLayer                = codegen.BuiltInLayer
+	BuiltInViewportIndex        = codegen.BuiltInViewportIndex
+	BuiltInTessLevelOuter       = codegen.BuiltInTessLevelOuter
+	BuiltInTessLevelInner       = codegen.BuiltInTessLevelInner
+	BuiltInTessCoord            = codegen.BuiltInTessCoord
+	BuiltInPatchVertices        = codegen.BuiltInPatchVertices
+	BuiltInFragCoord            = codegen.BuiltInFragCoord
+	BuiltInPointCoord           = codegen.BuiltInPointCoord
+	BuiltInFrontFacing          = codegen.BuiltInFrontFacing
+	BuiltInSampleID             = codegen.BuiltInSampleID
+	BuiltInSamplePosition       = codegen.BuiltInSamplePosition
+	BuiltInSampleMask           = codegen.BuiltInSampleMask
+	BuiltInFragDepth            = codegen.BuiltInFragDepth
+	BuiltInHelperInvocation     = codegen.BuiltInHelperInvocation
+	BuiltInNumWorkgroups        = codegen.BuiltInNumWorkgroups
+	BuiltInWorkgroupSize        = codegen.BuiltInWorkgroupSize
+	BuiltInWorkgroupID          = codegen.BuiltInWorkgroupID
+	BuiltInLocalInvocationID    = codegen.BuiltInLocalInvocationID
+	BuiltInGlobalInvocationID   = codegen.BuiltInGlobalInvocationID
+	BuiltInLocalInvocationIndex = codegen.BuiltInLocalInvocationIndex
+	BuiltInVertexIndex          = codegen.BuiltInVertexIndex
+	BuiltInInstanceIndex        = codegen.BuiltInInstanceIndex
+	BuiltInSubgroupSize         = codegen.BuiltInSubgroupSize
+	BuiltInSubgroupLocalInvID   = codegen.BuiltInSubgroupLocalInvID
+	BuiltInNumSubgroups         = codegen.BuiltInNumSubgroups
+	BuiltInSubgroupID           = codegen.BuiltInSubgroupID
+	BuiltInViewIndex            = codegen.BuiltInViewIndex
+	BuiltInBaryCoordKHR         = codegen.BuiltInBaryCoordKHR
 )
 
-// ImageFormat represents a SPIR-V image format.
-type ImageFormat uint32
+// --- Execution models ---
 
-// SPIR-V image format values (for OpTypeImage)
 const (
-	ImageFormatUnknown      ImageFormat = 0
-	ImageFormatRgba32f      ImageFormat = 1
-	ImageFormatRgba16f      ImageFormat = 2
-	ImageFormatR32f         ImageFormat = 3
-	ImageFormatRgba8        ImageFormat = 4
-	ImageFormatRgba8Snorm   ImageFormat = 5
-	ImageFormatRg32f        ImageFormat = 6
-	ImageFormatRg16f        ImageFormat = 7
-	ImageFormatR11fG11fB10f ImageFormat = 8
-	ImageFormatR16f         ImageFormat = 9
-	ImageFormatRgba16       ImageFormat = 10
-	ImageFormatRgb10A2      ImageFormat = 11
-	ImageFormatRg16         ImageFormat = 12
-	ImageFormatRg8          ImageFormat = 13
-	ImageFormatR16          ImageFormat = 14
-	ImageFormatR8           ImageFormat = 15
-	ImageFormatRgba16Snorm  ImageFormat = 16
-	ImageFormatRg16Snorm    ImageFormat = 17
-	ImageFormatRg8Snorm     ImageFormat = 18
-	ImageFormatR16Snorm     ImageFormat = 19
-	ImageFormatR8Snorm      ImageFormat = 20
-	ImageFormatRgba32i      ImageFormat = 21
-	ImageFormatRgba16i      ImageFormat = 22
-	ImageFormatRgba8i       ImageFormat = 23
-	ImageFormatR32i         ImageFormat = 24
-	ImageFormatRg32i        ImageFormat = 25
-	ImageFormatRg16i        ImageFormat = 26
-	ImageFormatRg8i         ImageFormat = 27
-	ImageFormatR16i         ImageFormat = 28
-	ImageFormatR8i          ImageFormat = 29
-	ImageFormatRgba32ui     ImageFormat = 30
-	ImageFormatRgba16ui     ImageFormat = 31
-	ImageFormatRgba8ui      ImageFormat = 32
-	ImageFormatR32ui        ImageFormat = 33
-	ImageFormatRgb10a2ui    ImageFormat = 34
-	ImageFormatRg32ui       ImageFormat = 35
-	ImageFormatRg16ui       ImageFormat = 36
-	ImageFormatRg8ui        ImageFormat = 37
-	ImageFormatR16ui        ImageFormat = 38
-	ImageFormatR8ui         ImageFormat = 39
-	ImageFormatR64ui        ImageFormat = 40
-	ImageFormatR64i         ImageFormat = 41
+	ExecutionModelVertex                 = codegen.ExecutionModelVertex
+	ExecutionModelTessellationControl    = codegen.ExecutionModelTessellationControl
+	ExecutionModelTessellationEvaluation = codegen.ExecutionModelTessellationEvaluation
+	ExecutionModelGeometry               = codegen.ExecutionModelGeometry
+	ExecutionModelFragment               = codegen.ExecutionModelFragment
+	ExecutionModelGLCompute              = codegen.ExecutionModelGLCompute
+	ExecutionModelKernel                 = codegen.ExecutionModelKernel
 )
 
-// GLSL.std.450 extended instruction set constants
+// --- Execution modes ---
+
 const (
-	GLSLstd450Round                 uint32 = 1
-	GLSLstd450RoundEven             uint32 = 2
-	GLSLstd450Trunc                 uint32 = 3
-	GLSLstd450FAbs                  uint32 = 4
-	GLSLstd450SAbs                  uint32 = 5
-	GLSLstd450FSign                 uint32 = 6
-	GLSLstd450SSign                 uint32 = 7
-	GLSLstd450Floor                 uint32 = 8
-	GLSLstd450Ceil                  uint32 = 9
-	GLSLstd450Fract                 uint32 = 10
-	GLSLstd450Radians               uint32 = 11
-	GLSLstd450Degrees               uint32 = 12
-	GLSLstd450Sin                   uint32 = 13
-	GLSLstd450Cos                   uint32 = 14
-	GLSLstd450Tan                   uint32 = 15
-	GLSLstd450Asin                  uint32 = 16
-	GLSLstd450Acos                  uint32 = 17
-	GLSLstd450Atan                  uint32 = 18
-	GLSLstd450Sinh                  uint32 = 19
-	GLSLstd450Cosh                  uint32 = 20
-	GLSLstd450Tanh                  uint32 = 21
-	GLSLstd450Asinh                 uint32 = 22
-	GLSLstd450Acosh                 uint32 = 23
-	GLSLstd450Atanh                 uint32 = 24
-	GLSLstd450Atan2                 uint32 = 25
-	GLSLstd450Pow                   uint32 = 26
-	GLSLstd450Exp                   uint32 = 27
-	GLSLstd450Log                   uint32 = 28
-	GLSLstd450Exp2                  uint32 = 29
-	GLSLstd450Log2                  uint32 = 30
-	GLSLstd450Sqrt                  uint32 = 31
-	GLSLstd450InverseSqrt           uint32 = 32
-	GLSLstd450Determinant           uint32 = 33
-	GLSLstd450MatrixInverse         uint32 = 34
-	GLSLstd450Modf                  uint32 = 35
-	GLSLstd450ModfStruct            uint32 = 36
-	GLSLstd450FMin                  uint32 = 37
-	GLSLstd450UMin                  uint32 = 38
-	GLSLstd450SMin                  uint32 = 39
-	GLSLstd450FMax                  uint32 = 40
-	GLSLstd450UMax                  uint32 = 41
-	GLSLstd450SMax                  uint32 = 42
-	GLSLstd450FClamp                uint32 = 43
-	GLSLstd450UClamp                uint32 = 44
-	GLSLstd450SClamp                uint32 = 45
-	GLSLstd450FMix                  uint32 = 46
-	GLSLstd450IMix                  uint32 = 47
-	GLSLstd450Step                  uint32 = 48
-	GLSLstd450SmoothStep            uint32 = 49
-	GLSLstd450Fma                   uint32 = 50
-	GLSLstd450Frexp                 uint32 = 51
-	GLSLstd450FrexpStruct           uint32 = 52
-	GLSLstd450Ldexp                 uint32 = 53
-	GLSLstd450PackSnorm4x8          uint32 = 54
-	GLSLstd450PackUnorm4x8          uint32 = 55
-	GLSLstd450PackSnorm2x16         uint32 = 56
-	GLSLstd450PackUnorm2x16         uint32 = 57
-	GLSLstd450PackHalf2x16          uint32 = 58
-	GLSLstd450PackDouble2x32        uint32 = 59
-	GLSLstd450UnpackSnorm2x16       uint32 = 60
-	GLSLstd450UnpackUnorm2x16       uint32 = 61
-	GLSLstd450UnpackHalf2x16        uint32 = 62
-	GLSLstd450UnpackSnorm4x8        uint32 = 63
-	GLSLstd450UnpackUnorm4x8        uint32 = 64
-	GLSLstd450UnpackDouble2x32      uint32 = 65
-	GLSLstd450Length                uint32 = 66
-	GLSLstd450Distance              uint32 = 67
-	GLSLstd450Cross                 uint32 = 68
-	GLSLstd450Normalize             uint32 = 69
-	GLSLstd450FaceForward           uint32 = 70
-	GLSLstd450Reflect               uint32 = 71
-	GLSLstd450Refract               uint32 = 72
-	GLSLstd450FindILsb              uint32 = 73
-	GLSLstd450FindSMsb              uint32 = 74
-	GLSLstd450FindUMsb              uint32 = 75
-	GLSLstd450InterpolateAtCentroid uint32 = 76
-	GLSLstd450InterpolateAtSample   uint32 = 77
-	GLSLstd450InterpolateAtOffset   uint32 = 78
-	GLSLstd450NMin                  uint32 = 79
-	GLSLstd450NMax                  uint32 = 80
-	GLSLstd450NClamp                uint32 = 81
+	ExecutionModeInvocations              = codegen.ExecutionModeInvocations
+	ExecutionModeSpacingEqual             = codegen.ExecutionModeSpacingEqual
+	ExecutionModeSpacingFractionalEven    = codegen.ExecutionModeSpacingFractionalEven
+	ExecutionModeSpacingFractionalOdd     = codegen.ExecutionModeSpacingFractionalOdd
+	ExecutionModeVertexOrderCw            = codegen.ExecutionModeVertexOrderCw
+	ExecutionModeVertexOrderCcw           = codegen.ExecutionModeVertexOrderCcw
+	ExecutionModePixelCenterInteger       = codegen.ExecutionModePixelCenterInteger
+	ExecutionModeOriginUpperLeft          = codegen.ExecutionModeOriginUpperLeft
+	ExecutionModeOriginLowerLeft          = codegen.ExecutionModeOriginLowerLeft
+	ExecutionModeEarlyFragmentTests       = codegen.ExecutionModeEarlyFragmentTests
+	ExecutionModePointMode                = codegen.ExecutionModePointMode
+	ExecutionModeXfb                      = codegen.ExecutionModeXfb
+	ExecutionModeDepthReplacing           = codegen.ExecutionModeDepthReplacing
+	ExecutionModeDepthGreater             = codegen.ExecutionModeDepthGreater
+	ExecutionModeDepthLess                = codegen.ExecutionModeDepthLess
+	ExecutionModeDepthUnchanged           = codegen.ExecutionModeDepthUnchanged
+	ExecutionModeLocalSize                = codegen.ExecutionModeLocalSize
+	ExecutionModeLocalSizeHint            = codegen.ExecutionModeLocalSizeHint
+	ExecutionModeInputPoints              = codegen.ExecutionModeInputPoints
+	ExecutionModeInputLines               = codegen.ExecutionModeInputLines
+	ExecutionModeInputLinesAdjacency      = codegen.ExecutionModeInputLinesAdjacency
+	ExecutionModeTriangles                = codegen.ExecutionModeTriangles
+	ExecutionModeInputTrianglesAdjacency  = codegen.ExecutionModeInputTrianglesAdjacency
+	ExecutionModeQuads                    = codegen.ExecutionModeQuads
+	ExecutionModeIsolines                 = codegen.ExecutionModeIsolines
+	ExecutionModeOutputVertices           = codegen.ExecutionModeOutputVertices
+	ExecutionModeOutputPoints             = codegen.ExecutionModeOutputPoints
+	ExecutionModeOutputLineStrip          = codegen.ExecutionModeOutputLineStrip
+	ExecutionModeOutputTriangleStrip      = codegen.ExecutionModeOutputTriangleStrip
+	ExecutionModeVecTypeHint              = codegen.ExecutionModeVecTypeHint
+	ExecutionModeContractionOff           = codegen.ExecutionModeContractionOff
+	ExecutionModeInitializer              = codegen.ExecutionModeInitializer
+	ExecutionModeFinalizer                = codegen.ExecutionModeFinalizer
+	ExecutionModeSubgroupSize             = codegen.ExecutionModeSubgroupSize
+	ExecutionModeSubgroupsPerWorkgroup    = codegen.ExecutionModeSubgroupsPerWorkgroup
+	ExecutionModeSubgroupsPerWorkgroupID  = codegen.ExecutionModeSubgroupsPerWorkgroupID
+	ExecutionModeLocalSizeID              = codegen.ExecutionModeLocalSizeID
+	ExecutionModeLocalSizeHintID          = codegen.ExecutionModeLocalSizeHintID
+	ExecutionModePostDepthCoverage        = codegen.ExecutionModePostDepthCoverage
+	ExecutionModeDenormPreserve           = codegen.ExecutionModeDenormPreserve
+	ExecutionModeDenormFlushToZero        = codegen.ExecutionModeDenormFlushToZero
+	ExecutionModeSignedZeroInfNanPreserve = codegen.ExecutionModeSignedZeroInfNanPreserve
+	ExecutionModeRoundingModeRTE          = codegen.ExecutionModeRoundingModeRTE
+	ExecutionModeRoundingModeRTZ          = codegen.ExecutionModeRoundingModeRTZ
+)
+
+// --- Storage classes ---
+
+const (
+	StorageClassUniformConstant         = codegen.StorageClassUniformConstant
+	StorageClassInput                   = codegen.StorageClassInput
+	StorageClassUniform                 = codegen.StorageClassUniform
+	StorageClassOutput                  = codegen.StorageClassOutput
+	StorageClassWorkgroup               = codegen.StorageClassWorkgroup
+	StorageClassCrossWorkgroup          = codegen.StorageClassCrossWorkgroup
+	StorageClassPrivate                 = codegen.StorageClassPrivate
+	StorageClassFunction                = codegen.StorageClassFunction
+	StorageClassGeneric                 = codegen.StorageClassGeneric
+	StorageClassPushConstant            = codegen.StorageClassPushConstant
+	StorageClassAtomicCounter           = codegen.StorageClassAtomicCounter
+	StorageClassImage                   = codegen.StorageClassImage
+	StorageClassStorageBuffer           = codegen.StorageClassStorageBuffer
+	StorageClassTaskPayloadWorkgroupEXT = codegen.StorageClassTaskPayloadWorkgroupEXT
+)
+
+// --- Addressing models ---
+
+const (
+	AddressingModelLogical    = codegen.AddressingModelLogical
+	AddressingModelPhysical32 = codegen.AddressingModelPhysical32
+	AddressingModelPhysical64 = codegen.AddressingModelPhysical64
+)
+
+// --- Memory models ---
+
+const (
+	MemoryModelSimple  = codegen.MemoryModelSimple
+	MemoryModelGLSL450 = codegen.MemoryModelGLSL450
+	MemoryModelOpenCL  = codegen.MemoryModelOpenCL
+	MemoryModelVulkan  = codegen.MemoryModelVulkan
+)
+
+// --- Function control ---
+
+const (
+	FunctionControlNone       = codegen.FunctionControlNone
+	FunctionControlInline     = codegen.FunctionControlInline
+	FunctionControlDontInline = codegen.FunctionControlDontInline
+	FunctionControlPure       = codegen.FunctionControlPure
+	FunctionControlConst      = codegen.FunctionControlConst
+)
+
+// --- Selection control ---
+
+const (
+	SelectionControlNone        = codegen.SelectionControlNone
+	SelectionControlFlatten     = codegen.SelectionControlFlatten
+	SelectionControlDontFlatten = codegen.SelectionControlDontFlatten
+)
+
+// --- Loop control ---
+
+const (
+	LoopControlNone               = codegen.LoopControlNone
+	LoopControlUnroll             = codegen.LoopControlUnroll
+	LoopControlDontUnroll         = codegen.LoopControlDontUnroll
+	LoopControlDependencyInfinite = codegen.LoopControlDependencyInfinite
+	LoopControlDependencyLength   = codegen.LoopControlDependencyLength
+	LoopControlMinIterations      = codegen.LoopControlMinIterations
+	LoopControlMaxIterations      = codegen.LoopControlMaxIterations
+	LoopControlIterationMultiple  = codegen.LoopControlIterationMultiple
+	LoopControlPeelCount          = codegen.LoopControlPeelCount
+	LoopControlPartialCount       = codegen.LoopControlPartialCount
+)
+
+// --- Image formats ---
+
+const (
+	ImageFormatUnknown      = codegen.ImageFormatUnknown
+	ImageFormatRgba32f      = codegen.ImageFormatRgba32f
+	ImageFormatRgba16f      = codegen.ImageFormatRgba16f
+	ImageFormatR32f         = codegen.ImageFormatR32f
+	ImageFormatRgba8        = codegen.ImageFormatRgba8
+	ImageFormatRgba8Snorm   = codegen.ImageFormatRgba8Snorm
+	ImageFormatRg32f        = codegen.ImageFormatRg32f
+	ImageFormatRg16f        = codegen.ImageFormatRg16f
+	ImageFormatR11fG11fB10f = codegen.ImageFormatR11fG11fB10f
+	ImageFormatR16f         = codegen.ImageFormatR16f
+	ImageFormatRgba16       = codegen.ImageFormatRgba16
+	ImageFormatRgb10A2      = codegen.ImageFormatRgb10A2
+	ImageFormatRg16         = codegen.ImageFormatRg16
+	ImageFormatRg8          = codegen.ImageFormatRg8
+	ImageFormatR16          = codegen.ImageFormatR16
+	ImageFormatR8           = codegen.ImageFormatR8
+	ImageFormatRgba16Snorm  = codegen.ImageFormatRgba16Snorm
+	ImageFormatRg16Snorm    = codegen.ImageFormatRg16Snorm
+	ImageFormatRg8Snorm     = codegen.ImageFormatRg8Snorm
+	ImageFormatR16Snorm     = codegen.ImageFormatR16Snorm
+	ImageFormatR8Snorm      = codegen.ImageFormatR8Snorm
+	ImageFormatRgba32i      = codegen.ImageFormatRgba32i
+	ImageFormatRgba16i      = codegen.ImageFormatRgba16i
+	ImageFormatRgba8i       = codegen.ImageFormatRgba8i
+	ImageFormatR32i         = codegen.ImageFormatR32i
+	ImageFormatRg32i        = codegen.ImageFormatRg32i
+	ImageFormatRg16i        = codegen.ImageFormatRg16i
+	ImageFormatRg8i         = codegen.ImageFormatRg8i
+	ImageFormatR16i         = codegen.ImageFormatR16i
+	ImageFormatR8i          = codegen.ImageFormatR8i
+	ImageFormatRgba32ui     = codegen.ImageFormatRgba32ui
+	ImageFormatRgba16ui     = codegen.ImageFormatRgba16ui
+	ImageFormatRgba8ui      = codegen.ImageFormatRgba8ui
+	ImageFormatR32ui        = codegen.ImageFormatR32ui
+	ImageFormatRgb10a2ui    = codegen.ImageFormatRgb10a2ui
+	ImageFormatRg32ui       = codegen.ImageFormatRg32ui
+	ImageFormatRg16ui       = codegen.ImageFormatRg16ui
+	ImageFormatRg8ui        = codegen.ImageFormatRg8ui
+	ImageFormatR16ui        = codegen.ImageFormatR16ui
+	ImageFormatR8ui         = codegen.ImageFormatR8ui
+	ImageFormatR64ui        = codegen.ImageFormatR64ui
+	ImageFormatR64i         = codegen.ImageFormatR64i
+)
+
+// --- GLSL.std.450 extended instruction set ---
+
+const (
+	GLSLstd450Round                 = codegen.GLSLstd450Round
+	GLSLstd450RoundEven             = codegen.GLSLstd450RoundEven
+	GLSLstd450Trunc                 = codegen.GLSLstd450Trunc
+	GLSLstd450FAbs                  = codegen.GLSLstd450FAbs
+	GLSLstd450SAbs                  = codegen.GLSLstd450SAbs
+	GLSLstd450FSign                 = codegen.GLSLstd450FSign
+	GLSLstd450SSign                 = codegen.GLSLstd450SSign
+	GLSLstd450Floor                 = codegen.GLSLstd450Floor
+	GLSLstd450Ceil                  = codegen.GLSLstd450Ceil
+	GLSLstd450Fract                 = codegen.GLSLstd450Fract
+	GLSLstd450Radians               = codegen.GLSLstd450Radians
+	GLSLstd450Degrees               = codegen.GLSLstd450Degrees
+	GLSLstd450Sin                   = codegen.GLSLstd450Sin
+	GLSLstd450Cos                   = codegen.GLSLstd450Cos
+	GLSLstd450Tan                   = codegen.GLSLstd450Tan
+	GLSLstd450Asin                  = codegen.GLSLstd450Asin
+	GLSLstd450Acos                  = codegen.GLSLstd450Acos
+	GLSLstd450Atan                  = codegen.GLSLstd450Atan
+	GLSLstd450Sinh                  = codegen.GLSLstd450Sinh
+	GLSLstd450Cosh                  = codegen.GLSLstd450Cosh
+	GLSLstd450Tanh                  = codegen.GLSLstd450Tanh
+	GLSLstd450Asinh                 = codegen.GLSLstd450Asinh
+	GLSLstd450Acosh                 = codegen.GLSLstd450Acosh
+	GLSLstd450Atanh                 = codegen.GLSLstd450Atanh
+	GLSLstd450Atan2                 = codegen.GLSLstd450Atan2
+	GLSLstd450Pow                   = codegen.GLSLstd450Pow
+	GLSLstd450Exp                   = codegen.GLSLstd450Exp
+	GLSLstd450Log                   = codegen.GLSLstd450Log
+	GLSLstd450Exp2                  = codegen.GLSLstd450Exp2
+	GLSLstd450Log2                  = codegen.GLSLstd450Log2
+	GLSLstd450Sqrt                  = codegen.GLSLstd450Sqrt
+	GLSLstd450InverseSqrt           = codegen.GLSLstd450InverseSqrt
+	GLSLstd450Determinant           = codegen.GLSLstd450Determinant
+	GLSLstd450MatrixInverse         = codegen.GLSLstd450MatrixInverse
+	GLSLstd450Modf                  = codegen.GLSLstd450Modf
+	GLSLstd450ModfStruct            = codegen.GLSLstd450ModfStruct
+	GLSLstd450FMin                  = codegen.GLSLstd450FMin
+	GLSLstd450UMin                  = codegen.GLSLstd450UMin
+	GLSLstd450SMin                  = codegen.GLSLstd450SMin
+	GLSLstd450FMax                  = codegen.GLSLstd450FMax
+	GLSLstd450UMax                  = codegen.GLSLstd450UMax
+	GLSLstd450SMax                  = codegen.GLSLstd450SMax
+	GLSLstd450FClamp                = codegen.GLSLstd450FClamp
+	GLSLstd450UClamp                = codegen.GLSLstd450UClamp
+	GLSLstd450SClamp                = codegen.GLSLstd450SClamp
+	GLSLstd450FMix                  = codegen.GLSLstd450FMix
+	GLSLstd450IMix                  = codegen.GLSLstd450IMix
+	GLSLstd450Step                  = codegen.GLSLstd450Step
+	GLSLstd450SmoothStep            = codegen.GLSLstd450SmoothStep
+	GLSLstd450Fma                   = codegen.GLSLstd450Fma
+	GLSLstd450Frexp                 = codegen.GLSLstd450Frexp
+	GLSLstd450FrexpStruct           = codegen.GLSLstd450FrexpStruct
+	GLSLstd450Ldexp                 = codegen.GLSLstd450Ldexp
+	GLSLstd450PackSnorm4x8          = codegen.GLSLstd450PackSnorm4x8
+	GLSLstd450PackUnorm4x8          = codegen.GLSLstd450PackUnorm4x8
+	GLSLstd450PackSnorm2x16         = codegen.GLSLstd450PackSnorm2x16
+	GLSLstd450PackUnorm2x16         = codegen.GLSLstd450PackUnorm2x16
+	GLSLstd450PackHalf2x16          = codegen.GLSLstd450PackHalf2x16
+	GLSLstd450PackDouble2x32        = codegen.GLSLstd450PackDouble2x32
+	GLSLstd450UnpackSnorm2x16       = codegen.GLSLstd450UnpackSnorm2x16
+	GLSLstd450UnpackUnorm2x16       = codegen.GLSLstd450UnpackUnorm2x16
+	GLSLstd450UnpackHalf2x16        = codegen.GLSLstd450UnpackHalf2x16
+	GLSLstd450UnpackSnorm4x8        = codegen.GLSLstd450UnpackSnorm4x8
+	GLSLstd450UnpackUnorm4x8        = codegen.GLSLstd450UnpackUnorm4x8
+	GLSLstd450UnpackDouble2x32      = codegen.GLSLstd450UnpackDouble2x32
+	GLSLstd450Length                = codegen.GLSLstd450Length
+	GLSLstd450Distance              = codegen.GLSLstd450Distance
+	GLSLstd450Cross                 = codegen.GLSLstd450Cross
+	GLSLstd450Normalize             = codegen.GLSLstd450Normalize
+	GLSLstd450FaceForward           = codegen.GLSLstd450FaceForward
+	GLSLstd450Reflect               = codegen.GLSLstd450Reflect
+	GLSLstd450Refract               = codegen.GLSLstd450Refract
+	GLSLstd450FindILsb              = codegen.GLSLstd450FindILsb
+	GLSLstd450FindSMsb              = codegen.GLSLstd450FindSMsb
+	GLSLstd450FindUMsb              = codegen.GLSLstd450FindUMsb
+	GLSLstd450InterpolateAtCentroid = codegen.GLSLstd450InterpolateAtCentroid
+	GLSLstd450InterpolateAtSample   = codegen.GLSLstd450InterpolateAtSample
+	GLSLstd450InterpolateAtOffset   = codegen.GLSLstd450InterpolateAtOffset
+	GLSLstd450NMin                  = codegen.GLSLstd450NMin
+	GLSLstd450NMax                  = codegen.GLSLstd450NMax
+	GLSLstd450NClamp                = codegen.GLSLstd450NClamp
 )
 
 // StorageFormatToImageFormat converts an IR storage format to a SPIR-V image format.
 func StorageFormatToImageFormat(format ir.StorageFormat) ImageFormat {
-	switch format {
-	// 8-bit formats
-	case ir.StorageFormatR8Unorm:
-		return ImageFormatR8
-	case ir.StorageFormatR8Snorm:
-		return ImageFormatR8Snorm
-	case ir.StorageFormatR8Uint:
-		return ImageFormatR8ui
-	case ir.StorageFormatR8Sint:
-		return ImageFormatR8i
-
-	// 16-bit formats
-	case ir.StorageFormatR16Uint:
-		return ImageFormatR16ui
-	case ir.StorageFormatR16Sint:
-		return ImageFormatR16i
-	case ir.StorageFormatR16Float:
-		return ImageFormatR16f
-	case ir.StorageFormatRg8Unorm:
-		return ImageFormatRg8
-	case ir.StorageFormatRg8Snorm:
-		return ImageFormatRg8Snorm
-	case ir.StorageFormatRg8Uint:
-		return ImageFormatRg8ui
-	case ir.StorageFormatRg8Sint:
-		return ImageFormatRg8i
-
-	// 32-bit formats
-	case ir.StorageFormatR32Uint:
-		return ImageFormatR32ui
-	case ir.StorageFormatR32Sint:
-		return ImageFormatR32i
-	case ir.StorageFormatR32Float:
-		return ImageFormatR32f
-	case ir.StorageFormatRg16Uint:
-		return ImageFormatRg16ui
-	case ir.StorageFormatRg16Sint:
-		return ImageFormatRg16i
-	case ir.StorageFormatRg16Float:
-		return ImageFormatRg16f
-	case ir.StorageFormatRgba8Unorm:
-		return ImageFormatRgba8
-	case ir.StorageFormatRgba8Snorm:
-		return ImageFormatRgba8Snorm
-	case ir.StorageFormatRgba8Uint:
-		return ImageFormatRgba8ui
-	case ir.StorageFormatRgba8Sint:
-		return ImageFormatRgba8i
-	case ir.StorageFormatBgra8Unorm:
-		return ImageFormatRgba8 // BGRA not directly supported, use RGBA
-
-	// Packed 32-bit formats
-	case ir.StorageFormatRgb10a2Uint:
-		return ImageFormatRgb10a2ui
-	case ir.StorageFormatRgb10a2Unorm:
-		return ImageFormatRgb10A2
-	case ir.StorageFormatRg11b10Ufloat:
-		return ImageFormatR11fG11fB10f
-
-	// 64-bit formats
-	case ir.StorageFormatRg32Uint:
-		return ImageFormatRg32ui
-	case ir.StorageFormatRg32Sint:
-		return ImageFormatRg32i
-	case ir.StorageFormatRg32Float:
-		return ImageFormatRg32f
-	case ir.StorageFormatRgba16Uint:
-		return ImageFormatRgba16ui
-	case ir.StorageFormatRgba16Sint:
-		return ImageFormatRgba16i
-	case ir.StorageFormatRgba16Float:
-		return ImageFormatRgba16f
-
-	// 128-bit formats
-	case ir.StorageFormatRgba32Uint:
-		return ImageFormatRgba32ui
-	case ir.StorageFormatRgba32Sint:
-		return ImageFormatRgba32i
-	case ir.StorageFormatRgba32Float:
-		return ImageFormatRgba32f
-
-	// Normalized 16-bit per channel formats
-	case ir.StorageFormatR16Unorm:
-		return ImageFormatR16
-	case ir.StorageFormatR16Snorm:
-		return ImageFormatR16Snorm
-	case ir.StorageFormatRg16Unorm:
-		return ImageFormatRg16
-	case ir.StorageFormatRg16Snorm:
-		return ImageFormatRg16Snorm
-	case ir.StorageFormatRgba16Unorm:
-		return ImageFormatRgba16
-	case ir.StorageFormatRgba16Snorm:
-		return ImageFormatRgba16Snorm
-
-	// 64-bit integer formats (SPV_EXT_shader_image_int64)
-	case ir.StorageFormatR64Uint:
-		return ImageFormatR64ui
-	case ir.StorageFormatR64Sint:
-		return ImageFormatR64i
-
-	default:
-		return ImageFormatUnknown
-	}
+	return codegen.StorageFormatToImageFormat(format)
 }
